Make Map and IssuesList components
